Ek flux boundary continuity (#5064)
Fixes #5049

Adds the ghost communication for the ek flux boundaries. However this only fixes the issue when the size of the ghost layer is two (n_ghost_layers=2). This is due to the boundary needs all its neighbors for the kernel to work properly and we need to run the kernel on the first ghost layer cells,  so that the fluxes into and out of the neighboring real cells has the correct value.